// src/api/types/BasicError.ts
// This file was auto-generated by Fern from our API Definition.

export interface BasicError {
    error?: BasicError.Error_ | undefined;
}

export namespace BasicError {
    export interface Error_ {
        /** Error message */
        message: string;
        /** Error code */
        code?: string | undefined;
    }
}