// src/api/resources/tts/client/Client.ts
export declare namespace TtsClient {
    export type Options = BaseClientOptions;

    export interface RequestOptions extends BaseRequestOptions {}
}

export class TtsClient {
    protected readonly _options: NormalizedClientOptionsWithAuth<TtsClient.Options>;

    constructor(options: TtsClient.Options = {}) {
        this._options = normalizeClientOptionsWithAuth(options);
    }

    /**
     * Streams generated speech audio for the provided text.
     *
     * @throws {@link Phonic.BadRequestError}
     * @throws {@link Phonic.UnauthorizedError}
     * @throws {@link Phonic.ForbiddenError}
     * @throws {@link Phonic.InternalServerError}
     */
    public stream(
        request: Phonic.StreamTtsRequest,
        requestOptions?: TtsClient.RequestOptions,
    ): core.HttpResponsePromise<core.BinaryResponse> {
        return core.HttpResponsePromise.fromPromise(this.__stream(request, requestOptions));
    }

    private async __stream(
        request: Phonic.StreamTtsRequest,
        requestOptions?: TtsClient.RequestOptions,
    ): Promise<core.WithRawResponse<core.BinaryResponse>> {
        const _authRequest: core.AuthRequest = await this._options.authProvider.getAuthRequest();
        const _headers: core.Fetcher.Args["headers"] = mergeHeaders(
            _authRequest.headers,
            this._options?.headers,
            requestOptions?.headers,
        );
        const _response = await (this._options.fetcher ?? core.fetcher)<core.BinaryResponse>({
            url: core.url.join(
                (await core.Supplier.get(this._options.baseUrl)) ??
                    ((await core.Supplier.get(this._options.environment)) ?? environments.PhonicEnvironment.Default)
                        .base,
                "tts/stream",
            ),
            method: "POST",
            headers: _headers,
            contentType: "application/json",
            queryParameters: requestOptions?.queryParams,
            requestType: "json",
            body: request,
            responseType: "binary-response",
            timeoutMs: (requestOptions?.timeoutInSeconds ?? this._options?.timeoutInSeconds ?? 60) * 1000,
            maxRetries: requestOptions?.maxRetries ?? this._options?.maxRetries,
            abortSignal: requestOptions?.abortSignal,
            fetchFn: this._options?.fetch,
            logging: this._options.logging,
        });
        if (_response.ok) {
            return { data: _response.body, rawResponse: _response.rawResponse };
        }

        if (_response.error.reason === "status-code") {
            switch (_response.error.statusCode) {
                case 400:
                    throw new Phonic.BadRequestError(_response.error.body as unknown, _response.rawResponse);
                case 401:
                    throw new Phonic.UnauthorizedError(
                        _response.error.body as Phonic.BasicError,
                        _response.rawResponse,
                    );
                case 403:
                    throw new Phonic.ForbiddenError(_response.error.body as unknown, _response.rawResponse);
                case 500:
                    throw new Phonic.InternalServerError(
                        _response.error.body as Phonic.BasicError,
                        _response.rawResponse,
                    );
                default:
                    throw new errors.PhonicError({
                        statusCode: _response.error.statusCode,
                        body: _response.error.body,
                        rawResponse: _response.rawResponse,
                    });
            }
        }

        return handleNonStatusCodeError(_response.error, _response.rawResponse, "POST", "/tts/stream");
    }
}